ドロップでfileを挿入できるようにした
