+++ features/outputs/types/editor.ts
@@ -28,3 +28,9 @@ export type HandleBeforeInputProps = EditorStateProps &
   SetEditorStateProps & {
     chars: string;
   };
+
+export type ToolbarButtonsProps = EditorStateProps & SetEditorStateProps;
+
+export type InsertImageToEditorProps = EditorStateProps & SetEditorStateProps & {
+  file: File;
+};
